+++ src/Services/AuthenticationService/AppyNox.Services.Authentication.WebAPI/Program.cs
@@ -207,7 +207,7 @@ if (app.Environment.IsDevelopment())
 
 app.UseRequestLocalization();
 
-app.UseCorrelationContext();
+app.UseNoxContext();
 
 app.UseNoxResponseWrapper(new NoxResponseWrapperOptions
 {