+++ src/Services/LicenseService/AppyNox.Services.License.WebAPI/Program.cs
@@ -194,7 +194,7 @@ if (!app.Environment.IsDevelopment())
     app.UseSwaggerUI();
 }
 
-app.UseCorrelationContext();
+app.UseNoxContext();
 
 app.UseNoxResponseWrapper(new NoxResponseWrapperOptions
 {
@@ -206,8 +206,6 @@ app.UseHttpsRedirection();
 app.UseAuthentication();
 
 app.UseAuthorization();
-
-app.UseUserIdContext();
 
 app.MapControllers();